style: move the save and load functions to their own module
started around 4:02 EST
